webapi/tct-geodeny-w3c-tests/geodeny/bdd/Geolocation_getCurrentPosition_operation_failed-manual.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Liu,Xin <xinx.liu@intel.com>

-->

<html>
  <head>
    <title>Geolocation Test:Geolocation_getCurrentPosition_operation_failed</title>
    <link rel="author" title="Intel" href="http://www.intel.com/" />
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#geolocation" />
    <meta name="flags" content="interact" />
    <meta name="assert" content="Check if getCurrentPosition() invokes onPositionErrorCallback() when the operation fails before the timeout expires" />
    <script src="../../resources/testharness.js" type="text/javascript"></script>
    <script src="../../resources/testharnessreport.js" type="text/javascript"></script>
  </head>
  <body>
    <h1>Browser settings do not share geographic location.</h1>
    <div id="log"></div>
    <script type="text/javascript">
        var t = async_test(document.title,{timeout:3000});
        setup({timeout:20000});
        function showGeolocation () {
            try {
                navigator.geolocation.getCurrentPosition(successCallback, errorCallback, { timeout : 6000});
            } catch (err) {
                t.step(function () {
                    assert_true(false,err.message);
                });
                t.done();
            }finally{
                setTimeout(function () {
                    t.done();
                }, 3000)
            }
        }
        function successCallback (position) {
            t.step(function () {
                assert_true(false,"This errorCallback should be called");
            });
            t.done();
        }
        function errorCallback (error) {
            var error = error.code;
            t.step(function () {
                assert_true(error !== null,"Get PositionErrorCallback object");
            });
            t.done();
        }
        showGeolocation();
  </script>
</body>
</html>
